Remove unused helper functions and relocate components
Deprecated `checkDecimalCount` and `formatNumber` helper functions were removed as they are no longer in use. Components were reorganized to follow a more consistent directory structure by relocating them to `components` with updated imports accordingly.
