+++ components/profile/bowCard/BowCardStyles.ts
