--- utils/helpers/formatNumber.ts
@@ -1,8 +1,0 @@
-/**
- * Remove comma and replace it with dot
- * @param number
- * @returns Formatted number
- */
-export function formatNumber(number: string) {
-  return number.replace(',', '.');
-}
