+++ utils/helpers/handleNumberChange.ts
@@ -1,4 +1,7 @@
-import { checkDecimalCount, formatNumber } from '@/utils';
+function checkDecimalCount(value: string, maxDecimals: number): boolean {
+  const decimalCount = (value.split('.')[1] || []).length;
+  return decimalCount < maxDecimals;
+}
 
 /**
  * Check decimal count in number input from the user
@@ -14,7 +17,7 @@ export function handleNumberChange(value: string, key: any, dispatch: any) {
     return;
   }
   if (!isNaN(parsedValue)) {
-    dispatch({ type: key, payload: formatNumber(value) });
+    dispatch({ type: key, payload: value.replace(',', '.') });
   } else {
     dispatch({ type: key, payload: '' });
   }
