+++ utils/index.ts
@@ -1,9 +1,7 @@
 export { Ballistics } from './Constants';
 export { default as getLocalStorage } from './LocalStorage/getLocalStorage';
 export { storeLocalStorage } from './LocalStorage/storeLocal';
-export { formatNumber } from './helpers/formatNumber';
 export { capitalizeFirstLetter } from './helpers/capitalizeFirstLetter';
 export { default as useBallisticsParams } from './hooks/useBallisticsParams';
 export { default as useCalculateMarks } from './hooks/useCalculateMarks';
-export { checkDecimalCount } from './helpers/checkDecimalCount';
 export { handleNumberChange } from './helpers/handleNumberChange';
